Got bullets working on both client- and server-side
