+++ Assets/Scripts/BulletPoolScript.cs
@@ -1,24 +1,19 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Networking;
 
-public class BulletScript : MonoBehaviour
+public class BulletPoolScript : NetworkBehaviour
 {
-    float lifetime;
-
     // Start is called before the first frame update
     void Start()
     {
-        lifetime = 1.0f;
+        
     }
 
     // Update is called once per frame
     void Update()
     {
-        lifetime -= Time.deltaTime;
-        if (lifetime < 0)
-        {
-            Destroy(this.gameObject);
-        }
+        
     }
 }
