+++ Assets/Scripts/BulletPoolScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 016f80f4695582448b6450a7afd017ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
